backend/requirements.txt
flask
google-genai
gunicorn
flask-cors
requests
python-dotenv
